Refactor API delegation service
