+++ src/features/apiDelegation/components/DelegationType.tsx
@@ -1,0 +1,4 @@
+export enum DelegationType {
+  Offered,
+  Received,
+}
