--- src/features/apiDelegation/components/LayoutState.tsx
@@ -1,4 +1,0 @@
-export enum LayoutState {
-  Offered,
-  Received,
-}
